requirements.txt
celery==4.2.1
itchat==1.3.10
gspread==3.1.0
oauth2client==4.1.3
dateparser==0.7.2
PyYAML==5.3